requirements.txt
absl-py==0.3.0
alabaster==0.7.11
astor==0.7.1
Babel==2.6.0
certifi==2018.4.16
chardet==3.0.4
cycler==0.10.0
dm-sonnet==1.23
docopt==0.6.2
docutils==0.14
gast==0.2.0
grpcio==1.13.0
idna==2.7
imagesize==1.0.0
Jinja2==2.10
jsonpickle==0.9.6
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
munch==2.3.2
networkx==1.8.1
numpy==1.15.0
packaging==17.1
Pillow==5.2.0
protobuf==3.6.0
py-cpuinfo==4.0.0
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
sacred==0.7.4
scipy==1.1.0
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.6
sphinxcontrib-websupport==1.1.0
tensorboard==1.9.0
tensorflow==1.9.0
termcolor==1.1.0
tqdm==4.24.0
urllib3==1.23
Werkzeug==0.14.1
wrapt==1.10.11
